Export RockSample from Domains (and clean up)
