+++ DecisionDomains.jl/Project.toml
@@ -3,3 +3,16 @@ uuid = "b3c2c752-e935-440e-b739-2b15001bb599"
 authors = ["Mel Krusniak <mel.krusniak@vanderbilt.edu>", "Jackson Wagner <jackson.wagner-1@colorado.edu>", "Kyle Wray <k.wray@northeastern.edu>", "Zachary Sunberg <zachary.sunberg@colorado.edu>", "Tyler Becker <tyler.becker-1@colorado.edu>"]
 version = "0.0.1"
 
+[deps]
+DecisionNetworks = "a1ae968e-b3ba-4546-81a6-f6e3d3a9391a"
+DecisionProblems = "b716e576-a1f9-415a-b454-baf909aed799"
+Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
+POMDPTools = "7588e00f-9cae-40de-98dc-e0c70c48cdd7"
+POMDPs = "a93abf59-7444-517b-a68a-c42f96afdd7d"
+RockSample = "de008ff0-c357-11e8-3329-7fe746fe836e"
+
+[compat]
+Distributions = "0.25.122"
+POMDPTools = "1.1.0"
+POMDPs = "1.0.0"
+RockSample = "0.2.1"
